removing old commanding mechanisms, improving cPU usage. Still needs documentation and getting rid of the nasty syncbuffering and runquickbuffering logic.
