+++ Unosquare.FFME.Common/Workers/BlockRenderingWorker.cs
@@ -1,5 +1,6 @@
 ﻿namespace Unosquare.FFME.Workers
 {
+    using Commands;
     using Primitives;
     using Shared;
     using System;
@@ -29,7 +30,7 @@
         /// <inheritdoc />
         public MediaEngine MediaCore { get; }
 
-        private CommandWorker Commands { get; }
+        private CommandManager Commands { get; }
 
         /// <inheritdoc />
         protected override void ExecuteCycleLogic(CancellationToken ct)
@@ -60,6 +61,9 @@
 
             try
             {
+                // TODO: wait for active seek command
+                Commands.WaitForSeekBlocks();
+
                 #region 2. Handle Block Rendering
 
                 // capture the wall clock for this cycle
